# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wayo API",
    "description": "Trip validation and owner-scoped draft storage. No itinerary generation yet.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/trips/validate": {
      "post": {
        "summary": "Validate Trip",
        "description": "Validate a draft without saving it or asserting itinerary feasibility.",
        "operationId": "validateTrip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripValidation"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips": {
      "post": {
        "tags": [
          "Trips"
        ],
        "summary": "Create Trip",
        "operationId": "createTrip",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTrip"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "200": {
            "description": "Previously created request replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTrip"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "List Trips",
        "operationId": "listTrips",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripList"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/v1/trips/{trip_id}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "summary": "Get Trip",
        "operationId": "getTrip",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTrip"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "put": {
        "tags": [
          "Trips"
        ],
        "summary": "Update Trip",
        "operationId": "updateTrip",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedTrip"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Trips"
        ],
        "summary": "Delete Trip",
        "operationId": "deleteTrip",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "expected_revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Expected Revision"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Anchor": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Label"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "latitude",
          "longitude"
        ],
        "title": "Anchor"
      },
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError"
      },
      "Budget": {
        "properties": {
          "amount_vnd": {
            "type": "integer",
            "maximum": 1000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amount Vnd"
          },
          "scope": {
            "type": "string",
            "enum": [
              "per_person",
              "group"
            ],
            "title": "Scope",
            "default": "per_person"
          },
          "mode": {
            "type": "string",
            "enum": [
              "soft",
              "hard"
            ],
            "title": "Mode",
            "default": "soft"
          },
          "currency": {
            "type": "string",
            "const": "VND",
            "title": "Currency",
            "default": "VND"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount_vnd"
        ],
        "title": "Budget"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail"
      },
...
      "FixedEvent": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Label"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "title": "Ends At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "starts_at",
          "ends_at"
        ],
        "title": "FixedEvent"
      },
...
      "InputNotice": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "InputNotice"
      },
...
      "Pace": {
        "type": "string",
        "enum": [
          "relaxed",
          "balanced",
          "active"
        ],
        "title": "Pace"
      },
...
      "SavedTrip": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "trip": {
            "$ref": "#/components/schemas/TripRequest"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "trip",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "SavedTrip"
      },
...
      "TripCreate": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "trip": {
            "$ref": "#/components/schemas/TripRequest"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "title",
          "trip"
        ],
        "title": "TripCreate"
      },
      "TripList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SavedTrip"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "offset",
          "limit"
        ],
        "title": "TripList"
      },
      "TripRequest": {
        "properties": {
          "destination_id": {
            "type": "string",
            "const": "da-lat",
            "title": "Destination Id",
            "default": "da-lat"
          },
          "origin": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Origin"
          },
          "arrival_at": {
            "type": "string",
            "format": "date-time",
            "title": "Arrival At"
          },
          "departure_at": {
            "type": "string",
            "format": "date-time",
            "title": "Departure At"
          },
          "timezone": {
            "type": "string",
            "const": "Asia/Ho_Chi_Minh",
            "title": "Timezone",
            "default": "Asia/Ho_Chi_Minh"
          },
          "people_count": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "People Count"
          },
          "group_type": {
            "type": "string",
            "enum": [
              "couple",
              "friends"
            ],
            "title": "Group Type"
          },
          "budget": {
            "$ref": "#/components/schemas/Budget"
          },
          "transport_mode": {
            "type": "string",
            "const": "driving",
            "title": "Transport Mode",
            "default": "driving"
          },
          "pace": {
            "$ref": "#/components/schemas/Pace",
            "default": "relaxed"
          },
          "anchor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Anchor"
              },
              {
                "type": "null"
              }
            ]
          },
          "preferences": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 20,
            "title": "Preferences"
          },
          "exclusions": {
            "items": {
              "type": "string",
              "maxLength": 120,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 20,
            "title": "Exclusions"
          },
          "fixed_events": {
            "items": {
              "$ref": "#/components/schemas/FixedEvent"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Fixed Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "origin",
          "arrival_at",
          "departure_at",
          "people_count",
          "group_type",
          "budget"
        ],
        "title": "TripRequest"
      },
      "TripUpdate": {
        "properties": {
          "expected_revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Revision"
          },
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "trip": {
            "$ref": "#/components/schemas/TripRequest"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "title",
          "trip"
        ],
        "title": "TripUpdate"
      },
      "TripValidation": {
        "properties": {
          "status": {
            "type": "string",
            "const": "input_valid",
            "title": "Status",
            "default": "input_valid"
          },
          "trip_days": {
            "type": "integer",
            "title": "Trip Days"
          },
          "total_budget_vnd": {
            "type": "integer",
            "title": "Total Budget Vnd"
          },
          "notices": {
            "items": {
              "$ref": "#/components/schemas/InputNotice"
            },
            "type": "array",
            "title": "Notices"
          },
          "trip": {
            "$ref": "#/components/schemas/TripRequest"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "trip_days",
          "total_budget_vnd",
          "notices",
          "trip"
        ],
        "title": "TripValidation"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}